src/test/javascript/spec/app/entities/transport/transport-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import TransportUpdateComponent from '@/entities/transport/transport-update.vue';
import TransportClass from '@/entities/transport/transport-update.component';
import TransportService from '@/entities/transport/transport.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Transport Management Update Component', () => {
    let wrapper: Wrapper<TransportClass>;
    let comp: TransportClass;
    let transportServiceStub: SinonStubbedInstance<TransportService>;

    beforeEach(() => {
      transportServiceStub = sinon.createStubInstance<TransportService>(TransportService);

      wrapper = shallowMount<TransportClass>(TransportUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          transportService: () => transportServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.transport = entity;
        transportServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(transportServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.transport = entity;
        transportServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(transportServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundTransport = { id: 123 };
        transportServiceStub.find.resolves(foundTransport);
        transportServiceStub.retrieve.resolves([foundTransport]);

        // WHEN
        comp.beforeRouteEnter({ params: { transportId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.transport).toBe(foundTransport);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
